refactor: move and split ImportService for future extensions
