+++ src/app/core/import/import.service.spec.ts
@@ -10,8 +10,6 @@ import {
   expectEntitiesToMatch,
 } from "../../utils/expect-entity-data.spec";
 import moment from "moment";
-import { RecurringActivity } from "../../child-dev-project/attendance/model/recurring-activity";
-import { ChildSchoolRelation } from "../../child-dev-project/children/model/childSchoolRelation";
 import { mockEntityMapper } from "../entity/entity-mapper/mock-entity-mapper-service";
 import { CoreTestingModule } from "../../utils/core-testing.module";
 import { EntityRegistry } from "../entity/database-entity.decorator";
@@ -69,6 +67,7 @@ describe("ImportService", () => {
       })
       entityRefs: string[];
     }
+
     spyOn(TestBed.inject(EntityRegistry), "get").and.callFake(
       (entityType: string) =>
         entityType === "ImportTestTarget" ? ImportTestTarget : TestEntity,
@@ -133,70 +132,22 @@ describe("ImportService", () => {
     );
   });
 
-  it("should link imported data to other entities", async () => {
-    const testEntities: Entity[] = [
-      createEntityOfType("Child", "1"),
-      createEntityOfType("Child", "2"),
-    ];
-    const activity = new RecurringActivity("3");
-    const entityMapper = TestBed.inject(EntityMapperService);
-    await entityMapper.save(activity);
-
-    const testImportSettings: ImportSettings = {
-      entityType: "Child",
-      columnMapping: undefined,
-      additionalActions: [
-        { type: "RecurringActivity", id: "RecurringActivity:3" },
-        { type: "School", id: "School:4" },
-      ],
-    };
-    await service.executeImport(testEntities, testImportSettings);
-
-    const createRelations = await entityMapper.loadType(ChildSchoolRelation);
-    const expectedRelations = [
-      { childId: "Child:1", schoolId: "School:4" },
-      { childId: "Child:2", schoolId: "School:4" },
-    ].map((e) => Object.assign(new ChildSchoolRelation(), e));
-    expectEntitiesToMatch(createRelations, expectedRelations, true);
-
-    expect(activity.participants).toEqual(["Child:1", "Child:2"]);
-  });
-
-  it("should allow to remove entities and links", async () => {
+  it("should allow to remove entities with undo", async () => {
     const importMeta = new ImportMetadata();
     importMeta.config = {
       entityType: "Child",
       columnMapping: undefined,
-      additionalActions: [
-        { type: "RecurringActivity", id: "3" },
-        { type: "School", id: "4" },
-      ],
     };
     importMeta.ids = ["Child:1", "Child:2"];
-    const relations = [
-      { childId: "1", schoolId: "4" },
-      { childId: "2", schoolId: "4" },
-      { childId: "3", schoolId: "4" }, // Other child same school -> keep
-      { childId: "2", schoolId: "3" }, // Imported child different school -> remove
-    ].map((e) => Object.assign(new ChildSchoolRelation(), e));
-    const activity = new RecurringActivity("3");
-    activity.participants = ["3", "2", "1"];
     const children = ["1", "2", "3"].map((id) =>
       createEntityOfType("Child", id),
     );
     const entityMapper = TestBed.inject(EntityMapperService);
-    await entityMapper.saveAll([
-      ...children,
-      ...relations,
-      activity,
-      importMeta,
-    ]);
+    await entityMapper.saveAll([...children, importMeta]);
 
     await service.undoImport(importMeta);
 
     await expectEntitiesToBeInDatabase([children[2]], false, true);
-    await expectEntitiesToBeInDatabase([relations[2]], false, true);
-    expect(activity.participants).toEqual(["3"]);
     await expectAsync(
       entityMapper.load(ImportMetadata, importMeta.getId()),
     ).toBeRejected();
